% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{boot_ardl_c}
\alias{boot_ardl_c}
\title{Code to generate data having a VECM-ARDL structure}
\arguments{
\item{r_in}{input residuals}

\item{GAMMAX}{short run parameter matrices column bound, first row in conditional form}

\item{A}{long-run parameter matrix, first row in conditional form}

\item{start_z}{data matrix of starting point}

\item{omegat}{parameter vector of the unlagged differences}

\item{interc}{vecm intercept, first element in conditional form}

\item{trend}{vecm trend, first element in conditional form}
}
\value{
A list containing the generated dataset
}
\description{
Code to generate data having a VECM-ARDL structure
}
\keyword{internal}
