% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_boot_ardl.R
\name{gen_boot_ardl}
\alias{gen_boot_ardl}
\title{Call to cpp function to generate data following a conditional ARDL/VECM model}
\usage{
gen_boot_ardl(r_in, GAMMAX, A, start_z, omegat, interc, trend)
}
\arguments{
\item{r_in}{matrix of residuals}

\item{GAMMAX}{short term parameters}

\item{A}{long term parameters}

\item{start_z}{initial observations}

\item{omegat}{parameter vector implied in conditioning}

\item{interc}{intercept vector}

\item{trend}{trend vector}
}
\value{
No return value, calls cpp function \code{boot_ardl_c}
}
\description{
Call to cpp function to generate data following a conditional ARDL/VECM model
}
\keyword{internal}
