% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ger_macro.R
\docType{data}
\name{ger_macro}
\alias{ger_macro}
\title{Investment, Income and Consumption dataset.}
\format{
A data frame with 92 rows and 4 variables:
\describe{
  \item{DATE}{Quarter}
  \item{INVEST}{Fixed investment (DM Billions)}
  \item{INCOME}{Disposable income (DM Billions)}
  \item{CONS}{Consumption expenditures (DM Billions)}
}
}
\source{
\url{http://www.jmulti.de/download/datasets/e1.dat}
}
\usage{
ger_macro
}
\description{
The data set contains quarterly, seasonally adjusted time series for 
West German fixed investment, disposable income, and consumption expenditures 
in billions of DM from 1960Q1 to 1982Q4. It was produced from file E1 of
 the data sets associated with Lutkepohl (2007). Originally obtained from Deutsche Bundesbank.
}
\keyword{datasets}
