% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ger_macro.R
\docType{data}
\name{ita_macro}
\alias{ita_macro}
\title{Export, Foreign Investment and GDP dataset.}
\format{
A data frame with 51 rows and 4 variables:
\describe{
  \item{YEAR}{From 1970 to 2020}
  \item{LEXP}{log Exports (Adjusted for DGP deflator)}
  \item{LFI}{log Foreign Direct Investment (Adjusted for DGP deflator)}
  \item{LGDP}{log GDP}
}
}
\source{
World Bank Database
}
\usage{
ita_macro
}
\description{
The data set contains yearly, time series for 
Italian log GDP, log foreign direct investment, and log exports, from 1970 to 2020.
Obtained from the World Bank Database.
}
\keyword{datasets}