% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smk_crit.R
\docType{data}
\name{smk_crit}
\alias{smk_crit}
\title{Critical values of the F-test on the independent variables in the conditional ARDL model.}
\format{
{A data frame with 144 rows and 17 variables}
\describe{
  \item{case}{case related to intercept and trend specification. Only I, III or V}
  \item{prob}{significance levels. 0.01, 0.025, 0.05 or 0.10}
  \item{num}{sample size, 30 to 80 and beyond for asymptotic critical values}
  \item{I0_x}{right threshold for accepting the null of zero coefficients in the Find test. x: 1 to 7}
  \item{I1_x}{left threshold for rejecting the null of zero coefficients in the Find test. x: 1 to }
}
}
\usage{
smk_crit
}
\description{
This internal data contains critical values of the Find test of Sam et al (2018), for several
sample sizes and  lag orders. Applicable only for cases I, III, and V.
Critical values at significance levels 1%, 2.5%, 5% and 10%.
}
\keyword{datasets}
