% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strip_formula_H0.R
\name{strip_formula_H0}
\alias{strip_formula_H0}
\title{Removes the variables imposed under the null from an ARDL formula object}
\usage{
strip_formula_H0(formula.ardl, H0, case, d)
}
\arguments{
\item{formula.ardl}{the ARDL formula object}

\item{H0}{the null hypothesis to be tested}

\item{case}{ARDL model case for the treatment of intercept and trend}

\item{d}{number of variables}
}
\value{
a formula object deprived of the terms set to zero by the null hypothesis
}
\description{
Removes the variables imposed under the null from an ARDL formula object
}
\keyword{internal}
