% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strip_formula_UC.R
\name{strip_formula_UC}
\alias{strip_formula_UC}
\title{Removes the unlagged differences imposed under the conditional ARDL model}
\usage{
strip_formula_UC(formula.ardl, case, d)
}
\arguments{
\item{formula.ardl}{the ARDL formula object}

\item{case}{ARDL model case for the treatment of intercept and trend}

\item{d}{number of variables}
}
\value{
a formula object deprived of the unlagged difference terms
}
\description{
Removes the unlagged differences imposed under the conditional ARDL model
}
\keyword{internal}
