% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ardl_to_lm.R
\name{ardl_to_lm}
\alias{ardl_to_lm}
\title{Convert ARDL to lm}
\usage{
ardl_to_lm(formula.ardl, case, d)
}
\arguments{
\item{formula.ardl}{the ARDL formula object}

\item{case}{ARDL model case for the treatment of intercept and trend}

\item{d}{number of variables}
}
\description{
Converts an ARDL formula to a lm formula
}
\keyword{internal}
